# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Olejos Backend API",
        "description": "API endpoints for the application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/invoices/generate-for-pickup/{pickup}": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Manually generate and submit an invoice for a completed pickup",
                "description": "Creates an invoice for the given pickup and immediately dispatches it to KSeF.",
                "operationId": "dd5205e0716d6c10443d03fbf2cb28e9",
                "parameters": [
                    {
                        "name": "pickup",
                        "in": "path",
                        "description": "Pickup ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invoice created successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "409": {
                        "description": "Conflict - Invoice exists"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}